+++ CHANGELOG.rst
@@ -27,6 +27,9 @@ Added
   sample direction.
 - Inverse pole figure colouring of orientations given a Laue symmetry and sample
   direction.
+- `from_axes_angles()` method to `Rotation` and `Orientation` as a shortcut to
+  `from_neo_euler()` for axis/angle pairs.
+- `Orientation` based classes now accept a `symmetry` argument upon initialisation.
 - Simple unit cell orientation plotting with `plot_unit_cell` for `Orientation`
   instances.
 
@@ -38,6 +41,9 @@ Changed
 - `Orientation` class methods `from_euler`, `from_matrix`, and `from_neo_euler` no longer 
   return the smallest angle orientation when a `symmetry` is given.
 - `CrystalMap.orientations` no longer returns smallest angle orientation.
+- The methods `flatten`, `reshape`, and `squeeze` have been overridden in
+  `Misorientation` based classes to maintain the initial symmetry of the returned
+  instance.
 
 Deprecated
 ----------